+++ composer.json
@@ -9,7 +9,9 @@
     "require-dev": {
         "johnkary/phpunit-speedtrap": "^4.0",
         "laminas/laminas-developer-tools": "*",
+        "laminas/laminas-modulemanager": "^2.9",
         "mockery/mockery": "^1.6",
+        "olcs/olcs-logging": "^5.0.0",
         "phpstan/phpstan": "^0.12.100",
         "phpunit/phpunit": "^9.6",
         "san/san-session-toolbar": "*"
@@ -27,7 +29,7 @@
   "autoload-dev": {
     "psr-4": {
       "OlcsTest\\": "test/Olcs/src",
-      "Permits\\": "test/Permits/src"
+      "PermitsTest\\": "test/Permits/src"
     }
   },
     "minimum-stability": "dev",
